% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{cal_confusion_matrix}
\alias{cal_confusion_matrix}
\title{Calculations from the confusion matrix}
\usage{
cal_confusion_matrix(mat_tab, cal = "acc")
}
\arguments{
\item{mat_tab}{A confusion matrix \code{table}}

\item{cal}{A \code{character} representing the value to be computed. "pre" calculates prevelance, "acc" calculates accuracy, "sen" calculates sensitivity, "spe" calculates "specificity, "fdr" calculates false discovery ratio, "dor" calculates diagnostic odds ratio}
}
\value{
A \code{character} of the computer value
}
\description{
This function calculates various values from the confusion matrix
}
\author{
Manisha Parajuli, Favour Onyido, Festus Attah
}
