% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{conf_mat}
\alias{conf_mat}
\title{Confusion matrix}
\usage{
conf_mat(x, y, cutoff = 0.5)
}
\arguments{
\item{x}{A \code{matrix} that holds the independent variables}

\item{y}{A \code{vector} representing the dependent variable}

\item{cutoff}{A \code{numeric} representing the threshold to classify the predicted probabilities}
}
\value{
A \code{table} containing the true positive, true negative, false positive, false negative values.
}
\description{
This function generates a confusion matrix for the predicted and actual values
}
\author{
Manisha Parajuli, Favour Onyido, Festus Attah
}
