% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{cost}
\alias{cost}
\title{Cost function}
\usage{
cost(beta, X, y)
}
\arguments{
\item{beta}{A \code{vector}}

\item{X}{A \code{matrix} that holds the independent variables}

\item{y}{A \code{vector} representing the dependent variable}
}
\value{
A \code{matrix} containing the following objects
}
\description{
This function computes the cost function that is to be optimized
}
\author{
Manisha Parajuli, Favour Onyido, Festus Attah
}
