% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{Optim_log}
\alias{Optim_log}
\title{Optim log}
\usage{
Optim_log(x, y)
}
\arguments{
\item{x}{A \code{matrix} that holds the independent variables}

\item{y}{A \code{vector} representing the dependent variable}
}
\value{
A \code{matrix} containing the intercept and coefficient
}
\description{
This function minimizes the cost function computed
}
\author{
Manisha Parajuli, Favour Onyido, Festus Attah
}
