% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{plot_predict}
\alias{plot_predict}
\title{Plot of the predicted probabilities vs the true data points}
\usage{
plot_predict(x, y)
}
\arguments{
\item{x}{A \code{matrix} that holds the independent variables}

\item{y}{A \code{vector} representing the dependent variable}
}
\value{
A \code{plot} of the predicted probabilities vrs. the true Data Points
}
\description{
This function plots the predicted probabilities vrs. the true Data Points
}
\author{
Manisha Parajuli, Favour Onyido, Festus Attah
}
